Revert 33996db8 - Instead, set the value up front before the watcher kicks in ...
By only updating the value when null, the assets field wasn't getting reset to an empty array.

The original issue 33996db8 was solving was that because `source` was getting changed within `mounted`, the watcher was running and updating the value to its default (usually null) and that's how it was being removed.