Revert `urls.enforce_trailing_slashes` config on SEO Pro side, as discussed.
We think most use cases revolve around use cases like Netlify/etc., so it's probable that the user then wants trailing slashes on all URLs, not just in SEO meta/sitemap. The fact that we were turning it on and off, could end up generating inaccurate stache. We can revisit later if this is needed.