Ensure hreflangs can be generated for taxonomy terms
Before this commit, a type error would be thrown due to the Entry typehint on the `hrefLangs` method.

This commit makes that method more generic so we can support terms too.

This commit also imports the `Entry` contract, rather than the implementation, otherwise it might fail when storing entries in the database.